Added comments for the iis 403 error
If entities are too big (like a huge node module posting a lot of messages while building) you will get a "System.Data.Services.Client.DataServiceClientException: The page was not displayed because the request entity is too large" error. If you want to get pass the 48k, you will need to apply this setting on IIS, since there is a know issue when ssl is enabled on a DataService.